Add files via upload
These figures were created as part of an interview process, where I was asked to do some analysis on an opportunity insights dataset of my choosing (https://opportunityinsights.org/data/). I decided to work with this dataset: The Association Between Income and Life Expectancy in the United States, 2001-2014. I used female life expectancy compared to the Shannon Diversity Index, as measured from the available data, as a baseline and compared the trends seen within this figure to other metrics like male life expectancy and whiteness to get a more complete picture of what’s driving the trends.